requirements.txt
aiohttp==3.7.3
amqp==5.0.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
Authlib==0.15.3
billiard==3.6.3.0
celery==5.0.0
celery-progress==0.0.14
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-stripe==2.4.1
Django==3.1.4
django-allauth==0.44.0
django-anymail==8.1
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-crontab==0.7.1
django-decouple==2.1
django-rest-auth==0.9.5
django-slack==5.16.0
django-timezone-field==4.1.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
itypes==1.2.0
Jinja2==2.11.2
jsonfield==3.1.0
kombu==5.0.2
MarkupSafe==1.1.1
multidict==5.1.0
oauthlib==3.1.0
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytracking==0.2.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
simple-salesforce==1.10.1
six==1.15.0
slackclient==2.9.3
sqlparse==0.4.1
stripe==2.54.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zipp==3.3.0
